requirements.txt
discord
pynacl
aiomysql
emoji
python-dotenv
requests
disputils==0.1.3
python-i18n[YAML]
psutil
py-cpuinfo
cryptography
